# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation My API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Get user transactions",
                "operationId": "getTransactions",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "query",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "transactions": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Transaction"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/transactions/{transactionId}": {
            "delete": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Delete a transaction",
                "operationId": "deleteTransaction",
                "parameters": [
                    {
                        "name": "transactionId",
                        "in": "path",
                        "description": "ID of the transaction",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction and transaction items deleted successfully"
                    }
                }
            }
        },
        "/api/transactions/{transactionId}/detail": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Get transaction details",
                "operationId": "transactionDetail",
                "parameters": [
                    {
                        "name": "transactionId",
                        "in": "path",
                        "description": "ID of the transaction",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "transaction_items": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "product_name": {
                                                        "type": "string"
                                                    },
                                                    "quantity": {
                                                        "type": "integer"
                                                    },
                                                    "price": {
                                                        "type": "number"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction items not found"
                    }
                }
            }
        },
        "/api/transactions/{transactionId}/cancel": {
            "put": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Cancel a transaction",
                "operationId": "cancelTransaction",
                "parameters": [
                    {
                        "name": "transactionId",
                        "in": "path",
                        "description": "ID of the transaction",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction cancelled successfully"
                    },
                    "404": {
                        "description": "Transaction not found"
                    }
                }
            }
        },
        "/api/transactions/{transactionId}/update-status": {
            "patch": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Update transaction status",
                "operationId": "updateStatus",
                "parameters": [
                    {
                        "name": "transactionId",
                        "in": "path",
                        "description": "ID of the transaction",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated status",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "on progress",
                                            "completed",
                                            "cancelled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction status updated successfully"
                    },
                    "404": {
                        "description": "Transaction not found"
                    }
                }
            }
        },
        "/api/checkout": {
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Checkout",
                "operationId": "checkout",
                "requestBody": {
                    "description": "Checkout details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "totalAmount",
                                    "cartItems"
                                ],
                                "properties": {
                                    "totalAmount": {
                                        "type": "number"
                                    },
                                    "cartItems": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "quantity",
                                                "price"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "type": "integer"
                                                },
                                                "price": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "transaction_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Checkout failed"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Transaction": {
                "title": "Transaction",
                "required": [
                    "id",
                    "user_id",
                    "total_amount",
                    "status",
                    "cancelled"
                ],
                "properties": {
                    "id": {
                        "description": "Transaction ID",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "User ID",
                        "type": "integer"
                    },
                    "total_amount": {
                        "description": "Total amount",
                        "type": "number",
                        "format": "float"
                    },
                    "status": {
                        "description": "Transaction status",
                        "type": "string"
                    },
                    "cancelled": {
                        "description": "Whether the transaction is cancelled",
                        "type": "boolean"
                    }
                },
                "type": "object"
            }
        }
    }
}